refactor(role): optimize role storage structure using map and split jsons
