+++ plugins/role/role_file_map.json
@@ -1,0 +1,42 @@
+{
+    猫娘: "cat_girl",
+    佛祖: "buddha",
+    英语翻译或修改: "en_translatorson",
+    写作助理: "writing_assistant",
+    语言输入优化: "language_optimize",
+    论文式回答: "essay_answer",
+    写作素材搜集: "writing_materials",
+    内容总结: "text_summarize",
+    格言书: "motto_book",
+    讲故事: "storyteller",
+    编剧: "screenwriter",
+    小说家: "novelist",
+    诗人: "poet",
+    新闻记者: "journalist",
+    论文学者: "scholar",
+    论文作家: "essay_writer",
+    同义词: "synonyms",
+    文本情绪分析: "emotion_analyze",
+    随机回复的疯子: "lunatic",
+    随机回复的醉鬼: "drunkard",
+    小红书风格: "xiaohongshu_style",
+    周报生成器: "weekly_report",
+    阴阳怪气语录生成器: "sarcasm_generator",
+    舔狗语录生成器: "simp_generator",
+    群聊取名: "group_name",
+    表情符号翻译器: "emoji_translator",
+    "AI 医生": "ai_doctor",
+    知识点阐述: "knowledge_explain",
+    辩手: "debater",
+    心理学家: "psychologist",
+    "IT 编程问题": "it_programming",
+    费曼学习法教练: "feynman_coach",
+    育儿帮手: "parenting_helper",
+    发言分析专家: "speech_analyze",
+    数据库专家: "database_expert",
+    自私基因: "selfish_gene",
+    智囊团: "think_tank",
+    算法竞赛专家: "algorithm_expert",
+    哲学家: "philosopher",
+    苏格拉底: "socrates",
+}
