+++ plugins/role/tag.json
@@ -1,0 +1,17 @@
+{
+    "tags": {
+    "favorite": "常用",
+    "mind": "思维",
+    "write": "写作",
+    "article": "文章",
+    "text": "文本",
+    "comments": "点评",
+    "code": "编程",
+    "life": "生活百科",
+    "interesting": "有趣",
+    "language": "语言",
+    "speech": "辩论",
+    "social": "社交",
+    "philosophy": "哲学"
+  }
+}
